loading reportViews earlier
